% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_100.R
\name{pace_100}
\alias{pace_100}
\title{Calculate race pace 50's depending on stroke and goal time for 100 yard or meter, specified by user.}
\usage{
pace_100(goal_time, stroke)
}
\arguments{
\item{goal_time}{The time that the swimmer wants to achieve in their race.}

\item{stroke}{The stroke the swimmer is swimming in the race.}
}
\value{
Print pace time.
}
\description{
Calculate race pace 50's depending on stroke and goal time for 100 yard or meter, specified by user.
}
\examples{
pace(55, "butterfly")
pace(60, "free")
pace(70, "breast")
}
