% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_200.R
\name{pace_200}
\alias{pace_200}
\title{Calculate race pace 50's depending on stroke and goal time for 200 yard or meter, specified by user.}
\usage{
pace_200(goal_time, stroke)
}
\arguments{
\item{goal_time}{The time that the swimmer wants to achieve in their race.}

\item{stroke}{The stroke the swimmer is swimming in the race.}
}
\value{
Print race pace times.
}
\description{
Calculate race pace 50's depending on stroke and goal time for 200 yard or meter, specified by user.
}
\examples{
pace_200(128, "breast")
pace_200(110, "freestyle")
}
